test: add SHM flow-control deadlock reproducer (currently skipped)
Documents a real architectural shortcoming: ShmServerTransport credits

HTTP/2 flow control only after the per-stream lpmAccumulator emits a

complete LPM (see handleMessage / handleMessageBuffer). When the

producer's send quota is smaller than a single MESSAGE the consumer

never sends a big-enough WindowUpdate to let the producer drain the

rest of the LPM, and the call deadlocks.

Production is currently masked from this by:

  - NewShmClientTransport hardcodes connSendQuota = maxWindowSize

    (2 GiB), so the producer never blocks on conn-level quota even

    for very large messages.

  - NewStream hardcodes streamSendQuota[id] = maxWindowSize, same.

  - NewShmServerTransport / NewStream pin inFlow.limit = maxWindowSize

    on the receive side.

  - grpc.WithInitialWindowSize is NOT plumbed through to the SHM

    transport, so users can't accidentally lower these quotas.

i.e. SHM transport currently runs with HTTP/2 flow control effectively

disabled; the ring buffer is the actual backpressure signal. Doug's

fair-comparison benchmark request (compare SHM against UDS / TCP with

equal HTTP/2 settings) will require either:

  - wiring WithInitialWindowSize through and fixing the deadlock

    (per-DATA-frame fc.onData accounting + producer chunked write), or

  - documenting that SHM uses ring backpressure and benchmarking

    against a TCP/UDS configuration tuned similarly.

Either path benefits from a pinned reproducer that fails fast and

dumps goroutine state on deadlock; this commit adds that test plus

the ConfigureShmFlowControlForBench / ResetShmFlowControlForBench

knobs the eventual fix will use.

Changes:

  - shm_flow_control.go: convert shmInitialWindowSize and

    shmWindowUpdateThreshold from const to var; add

    ConfigureShmFlowControlForBench(initialWindow) which clamps the

    threshold proportionally; add ResetShmFlowControlForBench for

    cleanup. NO behavioural change at defaults.

  - shm_client_transport.go / shm_server_transport.go: explicit

    type casts to keep build green after const -> var conversion.

  - shm_optimizations_test.go: same casts in existing tests.

  - shm_fc_small_window_test.go: TestShmSmallWindowMultiFrameMessage

    pinned to t.Skip; un-skip in the commit that lands the fix.

    The body of the test artificially clamps connSendQuota and

    streamSendQuota to 64 KiB after NewStream initialises them (to

    simulate the eventual WithInitialWindowSize plumbing) and

    asserts that a 256 KiB write completes.

Validation: go vet ./..., go test ./internal/transport/ (Windows,

36s). New test skips cleanly; pre-existing tests unaffected.

+++ internal/transport/shm_client_transport.go
@@ -229,7 +229,7 @@ func (t *ShmClientTransport) sendWindowUpdate(streamID uint32, delta uint32) {
 	t.sendQuotaMu.Lock()
 	if streamID == 0 {
 		t.pendingConnWU += delta
-		if t.pendingConnWU < shmWindowUpdateThreshold {
+		if t.pendingConnWU < uint32(shmWindowUpdateThreshold) {
 			t.sendQuotaMu.Unlock()
 			return
 		}
@@ -237,7 +237,7 @@ func (t *ShmClientTransport) sendWindowUpdate(streamID uint32, delta uint32) {
 		t.pendingConnWU = 0
 	} else {
 		t.pendingStreamWU[streamID] += delta
-		if t.pendingStreamWU[streamID] < shmWindowUpdateThreshold {
+		if t.pendingStreamWU[streamID] < uint32(shmWindowUpdateThreshold) {
 			t.sendQuotaMu.Unlock()
 			return
 		}
@@ -361,7 +361,7 @@ func NewShmClientTransport(segment *Segment, localAddr, remoteAddr net.Addr) (*S
 	// Initialize BDP estimation for dynamic flow control (RFC A73 Phase 5).
 	// SHM uses a much larger initial window (32MB) than HTTP/2 (64KB) because
 	// local memory has near-zero RTT and high bandwidth.
-	t.initialWindowSize = shmInitialWindowSize
+	t.initialWindowSize = int32(shmInitialWindowSize)
 	t.bdpEst = newShmBDPEstimator(uint32(shmInitialWindowSize), t.updateFlowControl)
 
 	max := segment.H.MaxStreams()
